Convert domain `AudioFeed` into `RssDocument'` to render
Ugh, it took a long time to decide on this separation of concerns and
where to put what. We'll see if it works out in the end…

Note: the stack resolver was changed to `lts-20.19` (GHC 9.2.7) because
I couldn't build the `rss-conduit` package with the previous `nightly`
resolver (probably, because it's already gone?!). Updating to the latest
`nightly` worked fine for the build, but HLS doesn't support GHC 9.4.5.